refactor(node): reorganise internal node messaging mod and remove redundant private APIs
- Fixing missing AE update to be sent to sibling new Elders upon split.
- Removing some node's public API which shouldn't be needed for any user of it (e.g. a UI).

BREAKING CHANGE: some APIs were removed from node's public interface.
